Controller node tasks to install k3s
This role installs k3s on the controller node and stores the specific
node token to add additional worker nodes to the cluster. This setup uses
the default k3s install script and shell piping which isnt ideal. Future
improvements would be to use the k3s binary, systemctl and services
similar to the k3s-ansible repo